+++ test/test_persistent_tasks.jl
@@ -1,0 +1,32 @@
+module TestPersistentTasks
+
+include("preamble.jl")
+using Base: PkgId, UUID
+using Pkg: TOML
+
+function getid(name)
+    path = joinpath(@__DIR__, "pkgs", "PersistentTasks", name)
+    if path ∉ LOAD_PATH
+        pushfirst!(LOAD_PATH, path)
+    end
+    prj = TOML.parsefile(joinpath(path, "Project.toml"))
+    return PkgId(UUID(prj["uuid"]), prj["name"])
+end
+
+
+@testset "PersistentTasks" begin
+    @test !Aqua.has_persistent_tasks(getid("TransientTask"))
+
+    result = Aqua.find_persistent_tasks_deps(getid("TransientTask"))
+    @test result == []
+
+    if Base.VERSION >= v"1.10-"
+        @test Aqua.has_persistent_tasks(getid("PersistentTask"))
+
+        result = Aqua.find_persistent_tasks_deps(getid("UsesBoth"))
+        @test result == ["PersistentTask"]
+    end
+    filter!(str -> !occursin("PersistentTasks", str), LOAD_PATH)
+end
+
+end